Tighten the behavior of corner-case entry creation (CWD and /).
Added proper tests and some additional checks in case of creating an entry with
an effective path of '.' or '/'.

Moved abs_path usage from Entry to where it was really needed (only in the case
of '.' in EntryFactory).

Improved some comments.

Did some refactoring of EntryFactory to make the code easier to understand.