Added time and date capabilities to the Formatter and LocaleEnv.
Formatter now knows how to add time and date information about entries into the
entry cell and slide page.  It asks the LocaleEnv to format the given time
according to the locale.
The LocaleEnv returns a string with question marks when not given a time. This
is useful for things such as lack of permissions to read the metadata or broken
symlinks.