Refactored Entry classes unit tests.
Changed the unit tests of the Entry class and its subclasses to use
Test::Class, in order to improve coverage and code reuse.

Removed the code which did an optional lazy stat() in
MMGal::Entry::creation_time(). From now on lack of stat argument on creation
means it was not possible to get one, and it will not be retried. Adjusted the
EntryFactory class and tests accordingly.

Added an embedded description to the test jpg file.