Dir: added time range support to creation_time, and ignoring lost+found.
lost+found will be ignored when enumerating directory entries. It is usually a
non-readable directory in the roots of filesystems.

creation_time() on a directory may now return two values when called in list
context, and the directory has more than one sub-element. In that case, the
first value is the timestamp of the oldest element, and the second - of the
newest one.

Added a copy of subdir/thumbnails/p.png needed for testing this functionality
properly.