Added MMGal::ImageInfo class and tests.
This is currently just a wrapper for Image::Info, but will become a generic
wrapper for other image info libraries, decoupling the rest of the code from
their implementation-specific intrinacies.

Its constructor is called "read", as it seems it is impossible to mock method
"new".

The creation_time method returns the best available creation time from whatever
is provided in EXIF tags.