Added Image::ExifTool support.
Split off the common image info module code into MMGal::ImageInfo::Base and
turned MMGal::ImageInfo into a pure factory module.

Added unit tests for this implementation (actually the tests are the same, but
we run them twice, after forcing a given implamentation.

Changed some tests and test data to run successfully with the new
implementation as default - this was required as Image::Info seems not to
support embedded date information in the png files, so the apparent dates
changed.

Added a note to the README files stating that Image::ExifTool is recommended.