Rename index.png and mmgal.css to start with ".mmgal-".
This way all mmgal-created files will start with ".mmgal-", apart from
index.html, which it is useful to have under this name, as it is the default
DirectoryIndex in Apache. It is possible to have any other name as the index,
but this way it is one thing less for the user to configure.